postgresql_api/postgresql_api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
python-dotenv==1.0.0